Create README.md
Rachael Fanifosi is passionate about problem-solving and delivering impactful results. She excels in leadership, people management, and driving projects to successful outcomes. 
Currently, Rachael is pursuing a career in cybersecurity, motivated by her interest in protecting digital assets.
If you are interested in collaborating in the field of cybersecurity, feel free to connect with her.